tianchi-enso-prediction/requirements.txt
certifi==2020.12.5
cftime==1.4.1
cycler==0.10.0
graphviz==0.16
h5netcdf==0.8.1
h5py==2.10.0
joblib==1.0.0
kiwisolver==1.3.1
matplotlib==3.3.4
netCDF4==1.5.5.1
# numpy==1.20.0
pandas==1.1.5
# Pillow==8.1.0
plotly==4.14.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
retrying==1.3.3
scikit-learn==0.24.1
scipy==1.4.1
six==1.15.0
threadpoolctl==2.1.0
tqdm==4.56.0
zipfile36==0.1.3
xarray==0.16.2
# lightgbm==3.1.1
pyarrow==3.0.0
tensorflow-gpu==2.4.1
# torch==1.4
global_land_mask==1.0.0
dtaidistance==2.2.5